StartupIQ_backend/requirements.txt
fastapi
uvicorn[standard]
motor
python-dotenv
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
python-multipart
scikit-learn
pandas
numpy
groq
